pyproject.toml
[tool.poetry]
name = "midi-drums"
version = "0.1.0"
description = ""
authors = ["Thiago F Pappacena <pappacena@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mido = "^1.3.2"
python-rtmidi = "^1.5.8"
pillow = "^10.4.0"
opencv-python-headless = "^4.10.0.84"
ollama = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
